+++ crates/libs/strings/src/hstring.rs
@@ -54,7 +54,7 @@ impl HSTRING {
     }
 
     /// Get the contents of this `HSTRING` as a OsString.
-    #[cfg(all(feature = "std", windows))]
+    #[cfg(feature = "std")]
     pub fn to_os_string(&self) -> std::ffi::OsString {
         std::os::windows::ffi::OsStringExt::from_wide(self.as_wide())
     }
@@ -154,14 +154,14 @@ impl From<&String> for HSTRING {
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl From<&std::path::Path> for HSTRING {
     fn from(value: &std::path::Path) -> Self {
         value.as_os_str().into()
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl From<&std::ffi::OsStr> for HSTRING {
     fn from(value: &std::ffi::OsStr) -> Self {
         unsafe {
@@ -174,14 +174,14 @@ impl From<&std::ffi::OsStr> for HSTRING {
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl From<std::ffi::OsString> for HSTRING {
     fn from(value: std::ffi::OsString) -> Self {
         value.as_os_str().into()
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl From<&std::ffi::OsString> for HSTRING {
     fn from(value: &std::ffi::OsString) -> Self {
         value.as_os_str().into()
@@ -286,28 +286,28 @@ impl PartialEq<&HSTRING> for String {
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<std::ffi::OsString> for HSTRING {
     fn eq(&self, other: &std::ffi::OsString) -> bool {
         *self == **other
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<std::ffi::OsString> for &HSTRING {
     fn eq(&self, other: &std::ffi::OsString) -> bool {
         **self == **other
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<&std::ffi::OsString> for HSTRING {
     fn eq(&self, other: &&std::ffi::OsString) -> bool {
         *self == ***other
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<std::ffi::OsStr> for HSTRING {
     fn eq(&self, other: &std::ffi::OsStr) -> bool {
         self.as_wide()
@@ -317,56 +317,56 @@ impl PartialEq<std::ffi::OsStr> for HSTRING {
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<std::ffi::OsStr> for &HSTRING {
     fn eq(&self, other: &std::ffi::OsStr) -> bool {
         **self == *other
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<&std::ffi::OsStr> for HSTRING {
     fn eq(&self, other: &&std::ffi::OsStr) -> bool {
         *self == **other
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<HSTRING> for std::ffi::OsStr {
     fn eq(&self, other: &HSTRING) -> bool {
         *other == *self
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<HSTRING> for &std::ffi::OsStr {
     fn eq(&self, other: &HSTRING) -> bool {
         *other == **self
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<&HSTRING> for std::ffi::OsStr {
     fn eq(&self, other: &&HSTRING) -> bool {
         **other == *self
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<HSTRING> for std::ffi::OsString {
     fn eq(&self, other: &HSTRING) -> bool {
         *other == **self
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<HSTRING> for &std::ffi::OsString {
     fn eq(&self, other: &HSTRING) -> bool {
         *other == ***self
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl PartialEq<&HSTRING> for std::ffi::OsString {
     fn eq(&self, other: &&HSTRING) -> bool {
         **other == **self
@@ -389,14 +389,14 @@ impl TryFrom<HSTRING> for String {
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl<'a> From<&'a HSTRING> for std::ffi::OsString {
     fn from(hstring: &HSTRING) -> Self {
         hstring.to_os_string()
     }
 }
 
-#[cfg(all(feature = "std", windows))]
+#[cfg(feature = "std")]
 impl From<HSTRING> for std::ffi::OsString {
     fn from(hstring: HSTRING) -> Self {
         Self::from(&hstring)
