+++ crates/libs/strings/src/lib.rs
@@ -2,6 +2,7 @@
 Learn more about Rust for Windows here: <https://github.com/microsoft/windows-rs>
 */
 
+#![cfg(windows)]
 #![allow(non_snake_case)]
 #![cfg_attr(
     windows_debugger_visualizer,
